Moved all cache eviction policies to CachePolicy.scala
